+++ flutter/packages/growerp/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.2.0
 
 - Now everything in a single repository.
+- upgraded the install , added first version of import and export.
 
 ## 0.1.0-dev.10